# source/mailingService/apiSpecification.yml
openapi: 3.0.3

info:
  title: Mailing-Service API
  description: This is the API-Specification for the Mailing-Service
  version: 1.0.0

paths:
  /sendMailAdded:
    post:
      summary: Interface for sending an email that a product was added
      description: Interface for sending an email with the given data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MailData'
      responses:
        '2XX':
          description: Response of the mailing service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONResponse'
        '4XX':
            description: Response of the mailing service
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/JSONResponse'
        '5XX':
            description: Response of the mailing service
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/JSONResponse'
  /sendMailDeleted:
    post:
      summary: Interface for sending an email that a product was deleted
      description: Interface for sending an email with the given data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MailData'
      responses:
        '2XX':
          description: Response of the mailing service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONResponse'
        '4XX':
            description: Response of the mailing service
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/JSONResponse'
        '5XX':
            description: Response of the mailing service
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/JSONResponse'
  /sendErrorMail:
    post:
      summary: Interface for sending an email that a error occured
      description: Interface for sending an email with the given data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MailData'
      responses:
        '2XX':
          description: Response of the mailing service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONResponse'
        '4XX':
            description: Response of the mailing service
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/JSONResponse'
        '5XX':
            description: Response of the mailing service
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/JSONResponse'
               
components:
  schemas:
    MailData:
      type: object
      required:
        - productId
        - productName
        - productPicture
      properties:
        productId:
          type: integer
          example: 123
        productName:
          type: string
          example: "Product XYZ"
        productPicture:
          type: string
          example: "Test-Anhang.png"
        errorMessage:
          type: string
          example: "The database service couldnt save the product"
    JSONResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: success
        message:
          type: string
          example: "Message sent successfully"